fix version check in modal header
