+++ src/lib/utils/hardware.ts
@@ -16,12 +16,4 @@ export const isPad = () => {
   return portraitWidthInches > 3.5
 }
 
-export const osMajorVersion = () => {
-  if (typeof (Platform.Version === "string")) {
-    return parseInt(Platform.Version as string, 10)
-  } else {
-    return Platform.Version as number
-  }
-}
-
 export const truncatedTextLimit = () => (isPad() ? 320 : 140)
